Add support for local files
