+++ fixtures/TEST_RELATIVE.html
@@ -1,0 +1,1 @@
+<a href="./TEST_RELATIVE_2.html">Foo</a>
