+++ fixtures/TEST_RELATIVE_2.html
@@ -1,0 +1,1 @@
+<a href="./TEST_RELATIVE_3.html">Bar</a>
