+++ fixtures/TEST_RELATIVE_3.html
@@ -1,0 +1,1 @@
+<a href="https://example.org">Example link</a>
