+++ lychee-lib/Cargo.toml
@@ -40,6 +40,7 @@ shellexpand = "2.1.0"
 tokio = { version = "1.6.0", features = ["full"] }
 typed-builder = "0.9.1"
 url = { version = "2.2.2", features = ["serde"] }
+log = "0.4.14"
 
 [dev-dependencies]
 doc-comment = "0.3.3"
