+++ lychee-lib/src/types/mod.rs
@@ -1,11 +1,15 @@
 #![allow(unreachable_pub)]
 
 mod error;
+mod file;
+mod input;
 mod request;
 mod response;
 mod status;
 
 pub use error::ErrorKind;
+pub use file::FileType;
+pub use input::{Input, InputContent};
 pub use request::Request;
 pub use response::{Response, ResponseBody};
 pub use status::Status;
